refactor: homework 패키지를 계층형 구조로 정리
docs/CONVENTIONS.md Part 1.2 의 [필수] 구조에 맞춘다.

  application  -> service
  presentation -> controller
  presentation/dto -> dto

헥사고날 잔재(application / infrastructure / presentation)를 걷어내고
classroom / member / openvidu 와 같은 구조가 되었다.

  homework
  ├── controller   2
  ├── service      2
  ├── repository   4
  ├── domain       6
  └── dto          5

테스트 164건 전부 통과

See also: #3
