+++ src/main/java/com/edu/edumeet/homework/dto/AssignmentCreateDTO.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.homework.presentation.dto;
+package com.edu.edumeet.homework.dto;
 
 import com.edu.edumeet.attachment.domain.Attachment;
 import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
