+++ src/main/java/com/edu/edumeet/homework/dto/StudentSubmissionStatusDTO.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.homework.presentation.dto;
+package com.edu.edumeet.homework.dto;
 
 import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
 import com.edu.edumeet.homework.domain.SubmissionStatus;
