+++ src/main/java/com/edu/edumeet/homework/dto/SubmissionCreateDTO.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.homework.presentation.dto;
+package com.edu.edumeet.homework.dto;
 
 import com.edu.edumeet.attachment.domain.Attachment;
 import io.swagger.v3.oas.annotations.media.Schema;
