+++ src/test/java/com/edu/edumeet/integration/homework/AssignmentListQueryCountTest.java
@@ -1,12 +1,12 @@
 package com.edu.edumeet.integration.homework;
 
-import com.edu.edumeet.homework.application.AssignmentService;
+import com.edu.edumeet.homework.service.AssignmentService;
 import com.edu.edumeet.homework.domain.SubmissionStatus;
 import com.edu.edumeet.homework.domain.AssignmentFileUpload;
 import com.edu.edumeet.homework.domain.Assignment;
 import com.edu.edumeet.homework.repository.AssignmentRepository;
 import com.edu.edumeet.homework.domain.StudentSubmissionStatus;
-import com.edu.edumeet.homework.presentation.dto.AssignmentDTO;
+import com.edu.edumeet.homework.dto.AssignmentDTO;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import lombok.extern.log4j.Log4j2;
