+++ src/test/java/com/edu/edumeet/integration/homework/CartesianProductTest.java
@@ -1,8 +1,8 @@
 package com.edu.edumeet.integration.homework;
 
-import com.edu.edumeet.homework.application.AssignmentService;
+import com.edu.edumeet.homework.service.AssignmentService;
 import com.edu.edumeet.homework.domain.SubmissionStatus;
-import com.edu.edumeet.homework.presentation.dto.AssignmentDTO;
+import com.edu.edumeet.homework.dto.AssignmentDTO;
 import com.edu.edumeet.homework.domain.AssignmentFileUpload;
 import com.edu.edumeet.homework.domain.Assignment;
 import com.edu.edumeet.homework.repository.AssignmentRepository;
