+++ src/test/java/com/edu/edumeet/unit/homework/application/AssignmentServiceImplTest.java
@@ -8,9 +8,9 @@ import com.edu.edumeet.homework.repository.AssignmentRepository;
 import com.edu.edumeet.member.domain.Member;
 import com.edu.edumeet.classroom.repository.ClassMemberRepository;
 import com.edu.edumeet.homework.domain.Assignment;
-import com.edu.edumeet.homework.application.AssignmentService;
-import com.edu.edumeet.homework.presentation.dto.AssignmentCreateDTO;
-import com.edu.edumeet.homework.presentation.dto.AssignmentDTO;
+import com.edu.edumeet.homework.service.AssignmentService;
+import com.edu.edumeet.homework.dto.AssignmentCreateDTO;
+import com.edu.edumeet.homework.dto.AssignmentDTO;
 import com.edu.edumeet.attachment.presentation.dto.AttachmentAdapter;
 import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
 import org.junit.jupiter.api.BeforeEach;
